feat: initial locale detection
